install  & create admin pages

+++ pages/member.php
@@ -11,7 +11,7 @@ $MustRegister = function () {
             <div class="col-md-3"></div>
             <div class="col-md-6">
                 <div class="alert alert-danger text-center my-3">
-                    You must login to view this page
+                    You have no permission to view this page
                 </div>
             </div>
             <div class="col-md-3"></div>
@@ -22,7 +22,7 @@ $MustRegister = function () {
 
 $Member = function() use ($title, $fetchMemberList, $editMember, $deleteMember, $MustRegister) {
     $title("Member");
-    if(!isset($_SESSION['member'])) return $MustRegister();
+    if(!isset($_SESSION['member']) || $_SESSION['status'] != 'admin') return $MustRegister();
     
     if(isset($_GET['delete'])) return $deleteMember();
     if(isset($_GET['edit'])) return $editMember();
